moving tag-action and snap-action to Harmony
